requirements.txt
google-api-python-client==2.128.0
gspread==6.1.0
pyarrow==12.0.1
mysql-connector-python==9.0.0
pandas==2.0.3
numpy==1.23.5
protobuf==3.20.3
pytz==2023.3
Scrapy==2.11.0
selenium==4.23.0
tabulate==0.9.0
Twisted==22.10.0
undetected_chromedriver==3.5.0
scrapy_xlsx==0.1.1
pymysql==1.1.0